add validators to other notes endpoints

+++ packages/api/src/modules/posts/controller/posts.controller.js
@@ -1,6 +1,6 @@
 const postsRepository = require("../repository/posts.repository");
 const postsService = require("../service/posts.service");
-const { validationResult } = require("express-validator/check");
+const { validationResult } = require("express-validator");
 
 exports.findNotes = async (req, res) => {
   const errors = validationResult(req);
@@ -13,34 +13,42 @@ exports.findNotes = async (req, res) => {
 };
 
 exports.findNotesBetweenDatesandKeyword = async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }
   const { page, perPage, email } = req.query;
-  const fromDate = new Date(req.params.fromDate);
-  const toDate = new Date(req.params.toDate);
+  const fromDate = req.params.fromDate;
+  const toDate = req.params.toDate;
   const keyword = req.params.keyword;
   toDate.setDate(toDate.getDate() + 1);
-  const pageAndLimit = {
-    page: parseInt(page, 10) || 1,
-    limit: parseInt(perPage, 10) || 10,
-  };
 
   const notes = await postsRepository.findNotesBetweenDatesandKeyword(
     fromDate,
     toDate,
     keyword,
     email,
-    pageAndLimit
+    { page, perPage }
   );
   res.send(notes);
 };
 
 exports.createPost = async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }
   const { email } = req.query;
   const { text } = req.body;
-  const newPost = await postsRepository.createPost(text, email, res);
+  const newPost = await postsRepository.createPost(text, email);
   res.status(201).send(newPost);
 };
 
 exports.editPost = async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }
   const { email } = req.query;
   const { text } = req.body;
   const id = req.params.id;
@@ -49,6 +57,10 @@ exports.editPost = async (req, res) => {
 };
 
 exports.deletePost = async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }
   const { email } = req.query;
   const id = req.params.id;
   const isUpdated = await postsRepository.deletePost(email, id);
